added requested changes
